requirements.txt

python-dotenv
google-generativeai
google-adk
PyPDF2
reportlab
presidio-analyzer
cryptography
spacy
